requirements.txt
numpy
requests
h5py
spacy
openpyxl
tqdm
pyyaml
simplejson
tensorflow==1.12